requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# LLM
anthropic==0.40.0

# Caching
cachetools==5.5.0

# Patch application
unidiff==0.7.5

# Docker SDK (sandbox)
docker==7.1.0

# YAML
pyyaml==6.0.2

# Observability
structlog==24.4.0
prometheus-client==0.21.1
opentelemetry-sdk==1.29.0
opentelemetry-api==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0

# Levenshtein distance (NameError typo detection)
python-Levenshtein==0.26.1

# HTTP client for tests
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-timeout==2.3.1

# Linting/type (dev)
ruff==0.8.4
mypy==1.13.0
slowapi==0.1.9
pydantic-settings==2.7.0
pydantic-settings==2.7.0
email-validator==2.2.0
